Use modifyRequest lifecycle hook for auth scheme
This changes gives other interceptors the opportunity to update auth scheme
related execution attributes such as IDENTITY_PROVIDERS before the auth scheme
is actually resolved. This is necessary because the auth scheme should not be
changed after it has been resolved as intermediate interceptors could be using
the resolved auth scheme, such as the accound ID in endpoint resolution.